Integrated CRUD Catalouge

+++ routes/tailor.js
@@ -13,6 +13,8 @@ const { v4: uuidv4 } = require('uuid');
 const bcrypt = require('bcryptjs');
 const passport = require('passport');
 const ensureAuthenticated = require('../helpers/auth');
+const Sequelize = require('sequelize');
+const Op = Sequelize.Op;
 
 
 var urlencodedParser = bodyParser.urlencoded({ extended: false });
@@ -22,11 +24,26 @@ function isNumeric(value) {
 }
 
 router.get('/addproduct', ensureAuthenticated, (req, res) => {
-	res.render('tailor/addproduct', { title: "Add product" });
+	var user_status = "customer";
+	if (typeof req.user != "undefined") {
+		user_status = res.locals.user.usertype;
+	}
+
+	if (user_status == "tailor") {
+		res.render('tailor/addproduct', { title: "Add product" });
+	}
+	else {
+		alertMessage(res, 'danger', 'Unauthorised access! Register as a tailor to add product.', 'fas fa-exclamation-triangle', true);
+		res.redirect('../hometailor');
+	}
 });
 
 // Add Product - POST
 router.post('/addproduct', ensureAuthenticated, urlencodedParser, (req, res) => {
+	if (typeof req.user != "undefined") {
+		var shopname = res.locals.user.shopname;
+	}
+
 	let errors = [];
 	let { name, price, discount, description, question, q1category } = req.body;
 
@@ -88,7 +105,7 @@ router.post('/addproduct', ensureAuthenticated, urlencodedParser, (req, res) => 
 
 		Catalouge
 			.create({
-				storename: 'Ah Tong Tailor',
+				storename: shopname,
 				name: name,
 				price: price,
 				image: newid + '.' + filetype,
@@ -150,41 +167,52 @@ router.post('/addproduct', ensureAuthenticated, urlencodedParser, (req, res) => 
 });
 
 // Update Product - GET
-router.get('/editproduct/:id', (req, res) => {
+router.get('/editproduct/:id', ensureAuthenticated, (req, res) => {
+	if (typeof req.user != "undefined") {
+		var shopname = res.locals.user.shopname;
+	}
+
 	Catalouge.findOne({
 		where: { id: req.params.id },
 		raw: true
 	})
 		.then(pdetails => {
 			console.log(pdetails);
-			if (pdetails) {
-				if (pdetails.customcat == "radiobtn") {
-					Productchoices.findAll({
-						where: { catalougeId: req.params.id },
-						raw: true
-					})
-						.then(pchoices => {
-							res.render('tailor/editproduct', {
-								title: "Update product",
-								pdetails: pdetails,
-								pchoices: pchoices
-							});
-						})
-						.catch(err => {
-							console.error('Unable to connect to the database:', err);
-						});
-
-				}
-				else {
-					res.render('tailor/editproduct', {
-						title: "Update product",
-						pdetails: pdetails
-					});
-				}
+			if (pdetails.storename != shopname) {
+				alertMessage(res, 'danger', 'You shall not pass!', 'fas fa-exclamation-triangle', true);
+				res.redirect('/viewshops');
 			}
 			else {
-				return res.redirect('/404');
+				if (pdetails) {
+					if (pdetails.customcat == "radiobtn") {
+						Productchoices.findAll({
+							where: { catalougeId: req.params.id },
+							raw: true
+						})
+							.then(pchoices => {
+								res.render('tailor/editproduct', {
+									title: "Update product",
+									pdetails: pdetails,
+									pchoices: pchoices
+								});
+							})
+							.catch(err => {
+								console.error('Unable to connect to the database:', err);
+							});
+
+					}
+					else {
+						res.render('tailor/editproduct', {
+							title: "Update product",
+							pdetails: pdetails
+						});
+					}
+				}
+				else {
+					return res.redirect('/404');
+				}
 			}
+
 		})
 		.catch(err => {
 			console.error('Unable to connect to the database:', err);
@@ -192,9 +220,13 @@ router.get('/editproduct/:id', (req, res) => {
 });
 
 // Update Product - PUT
-router.put('/editproduct/:id', urlencodedParser, (req, res) => {
+router.put('/editproduct/:id', ensureAuthenticated, urlencodedParser, (req, res) => {
 	let errors = [];
 	let { name, price, discount, description, question, q1category, imageLink } = req.body;
+
+	if (typeof req.user != "undefined") {
+		var shopname = res.locals.user.shopname;
+	}
 
 	// Validation
 	if (isNumeric(price) == false) {
@@ -295,19 +327,19 @@ router.put('/editproduct/:id', urlencodedParser, (req, res) => {
 
 			// Image Upload
 			var newFileName = newid;
-			console.log("./public/uploads/products/"+imageLink);
-			fs.unlink("./public/uploads/products/"+imageLink, (err) => {
+			console.log("./public/uploads/products/" + imageLink);
+			fs.unlink("./public/uploads/products/" + imageLink, (err) => {
 				if (err) {
-					console.log("failed to delete local image:"+err);
+					console.log("failed to delete local image:" + err);
 				} else {
-					console.log('successfully deleted local image');                                
+					console.log('successfully deleted local image');
 				}
 			});
 			file.mv('./public/uploads/products/' + filename, function (err) {
 				if (err) {
 					res.send(err);
 				}
-				else{
+				else {
 					fs.rename('./public/uploads/products/' + filename, './public/uploads/products/' + newFileName, function (err) {
 						if (err) console.log('ERROR: ' + err);
 					});
@@ -316,7 +348,7 @@ router.put('/editproduct/:id', urlencodedParser, (req, res) => {
 		}
 
 		Catalouge.update({
-			storename: 'Ah Tong Tailor',
+			storename: shopname,
 			name: name,
 			price: price,
 			image: newid,
@@ -326,12 +358,12 @@ router.put('/editproduct/:id', urlencodedParser, (req, res) => {
 			customcat: q1category
 		}, {
 			where: {
-				id: req.params.id
+				[Op.and]: [{id: req.params.id}, {storename: shopname}]
 			}
 		}).then(() => {
 
 			if (req.files) {
-				
+
 				// Page gets changed overly fast, such that page is loaded before img changes are made
 			}
 			alertMessage(res, 'success', 'Updated product successfully!', 'fas fa-check-circle', true);
@@ -350,13 +382,16 @@ router.put('/editproduct/:id', urlencodedParser, (req, res) => {
 	}
 });
 
-router.get('/deleteProduct/:id', (req, res) => {
+router.get('/deleteProduct/:id', ensureAuthenticated, (req, res) => {
+	if (typeof req.user != "undefined") {
+		var shopname = res.locals.user.shopname;
+	}
 	Catalouge.findOne({
 		where: {
 			id: req.params.id
 		}
 	}).then((pdetails) => {
-		if (pdetails != null) {
+		if (pdetails != null && pdetails.storename == shopname) {
 			if (pdetails.customcat == "radiobtn") {
 				Productchoices.destroy({
 					where: {
@@ -364,11 +399,11 @@ router.get('/deleteProduct/:id', (req, res) => {
 					}
 				});
 			}
-			fs.unlink("./public/uploads/products/"+pdetails.image, (err) => {
+			fs.unlink("./public/uploads/products/" + pdetails.image, (err) => {
 				if (err) {
-					console.log("failed to delete local image:"+err);
+					console.log("failed to delete local image:" + err);
 				} else {
-					console.log('successfully deleted local image');                                
+					console.log('successfully deleted local image');
 				}
 			});
 			Catalouge.destroy({
@@ -380,7 +415,11 @@ router.get('/deleteProduct/:id', (req, res) => {
 					alertMessage(res, 'info', 'Successfully deleted item.', 'far fa-trash-alt', true);
 					res.redirect('/viewshops/' + pdetails.storename);
 				})
-		};
+		}
+		else {
+			alertMessage(res, 'danger', 'Do you have a badge????', 'fas fa-exclamation-triangle', true);
+			res.redirect('/viewshops/' + pdetails.storename);
+		}
 	})
 })
 
@@ -395,10 +434,10 @@ router.post('/login', (req, res, next) => {
 		failureFlash: 'Invalid username or password.',
 		userProperty: res.user
 	})
-	(req, res, next);
+		(req, res, next);
 });
 
-function onSuccess(response){
+function onSuccess(response) {
 	return '/hometailor'
 }
 
@@ -428,7 +467,7 @@ router.post('/tailoregister', (req, res) => {
 		req.body.email,
 		req.body.phoneno,
 		req.body.usertype
-		);
+	);
 
 	// Checks if both passwords entered are the same
 	if (req.body.password !== req.body.password2) {
@@ -485,7 +524,7 @@ router.post('/tailoregister', (req, res) => {
 						bcrypt.hash(password, salt, (err, hash) => {
 							if (err) throw err;
 							password = hash;
-							User.create({ shopname, username, password, address1, address2, city, postalcode, email, phoneno,usertype:'tailor' })
+							User.create({ shopname, username, password, address1, address2, city, postalcode, email, phoneno, usertype: 'tailor' })
 								.then(user => {
 									alertMessage(res, 'success', user.username + ' Please proceed to login', 'fas fa-sign-in-alt', true);
 									res.redirect('tailoregcomplete');
@@ -501,7 +540,7 @@ router.post('/tailoregister', (req, res) => {
 });
 
 // tailor: account page 
-router.get('/tailoraccount/:id', ensureAuthenticated,(req, res) => {
+router.get('/tailoraccount/:id', ensureAuthenticated, (req, res) => {
 	User.findOne({
 		where: {
 			id: req.params.id
@@ -510,7 +549,7 @@ router.get('/tailoraccount/:id', ensureAuthenticated,(req, res) => {
 	}).then((Tailor) => {
 		console.log(Tailor);
 		if (req.params.id === Tailor.id) {
-			res.render('tailor/tailoracct', { 
+			res.render('tailor/tailoracct', {
 				User: Tailor
 			});
 		} else {
@@ -545,8 +584,8 @@ router.put('/tailoraccount/:id', ensureAuthenticated, (req, res) => {
 			id: req.params.id
 		}
 	}).then(() => {
-		alertMessage(res, 'success','Account has been updated successfully!', 'fas fa-sign-in-alt', true);
-		res.redirect('/tailor/tailoraccount/'+req.params.id);
+		alertMessage(res, 'success', 'Account has been updated successfully!', 'fas fa-sign-in-alt', true);
+		res.redirect('/tailor/tailoraccount/' + req.params.id);
 	}).catch(err => console.log(err));
 });
 module.exports = router;
